+++ src/polyfill.ts
@@ -1,0 +1,27 @@
+declare global {
+  // deno-lint-ignore no-explicit-any
+  interface ReadableStream<R = any> {
+    [Symbol.asyncIterator](options?: {
+      preventCancel?: boolean;
+    }): AsyncIterableIterator<R>;
+  }
+}
+
+// polyfill for ReadableStream.prototype[Symbol.asyncIterator]
+// https://bugs.chromium.org/p/chromium/issues/detail?id=929585#c10
+if (typeof ReadableStream.prototype[Symbol.asyncIterator] !== "function") {
+  ReadableStream.prototype[Symbol.asyncIterator] = async function* () {
+    const reader = this.getReader();
+    try {
+      while (true) {
+        const { done, value } = await reader.read();
+        if (done) return;
+        yield value;
+      }
+    } finally {
+      reader.releaseLock();
+    }
+  };
+}
+
+export {};
